refactor(backend): modularize api slices and router builders
