+++ docs/specs/9vau7-backend-structure-dual-pr-followup/SPEC.md
@@ -2,7 +2,7 @@
 
 ## 状态
 
-- Status: 待实现
+- Status: 已完成
 - Created: 2026-04-12
 - Last: 2026-04-12
 
@@ -75,9 +75,9 @@
 
 ## 实现里程碑（Milestones / Delivery checklist）
 
-- [ ] M1: 建立双 PR follow-up spec 并确认基线对齐 `main`
-- [ ] M2: 完成 PR1 (`proxy + upstream_accounts/routing`) 的结构收敛、验证、PR、合并与 cleanup
-- [ ] M3: 完成 PR2 (`api + router builders + api-read smoke`) 的结构收敛、验证、PR、合并与 cleanup
+- [x] M1: 建立双 PR follow-up spec 并确认基线对齐 `main`
+- [x] M2: 完成 PR1 (`proxy + upstream_accounts/routing`) 的结构收敛、验证、PR、合并与 cleanup
+- [x] M3: 完成 PR2 (`api + router builders + api-read smoke`) 的结构收敛、验证、PR、合并与 cleanup
 
 ## 方案概述（Approach, high-level）
 
@@ -94,6 +94,8 @@
 ## 变更记录（Change log）
 
 - 2026-04-12: 创建双 PR 后端结构债 fast-track spec，冻结 PR1/PR2 范围与 merge+cleanup 终点。
+- 2026-04-12: PR1 已完成 `proxy + upstream_accounts/routing` 真模块化并合入 `main`，本地 `cargo fmt/check/test` 与 shared-testbox `proxy-parallel/raw` smoke 通过。
+- 2026-04-12: PR2 已完成 `api + hourly_rollups router builders` 结构收敛，新增 `scripts/shared-testbox-api-read-smoke`，本地 `cargo fmt/check/test` 与 shared-testbox `api-read/proxy-parallel/raw` smoke 通过。
 
 ## 参考（References）
