+++ src/api/slices/mod.rs
@@ -1,0 +1,12 @@
+use crate::*;
+
+mod internal {
+    use super::*;
+
+    include!("invocations_and_summary.rs");
+    include!("error_distribution_and_sse.rs");
+    include!("settings_models_and_cache.rs");
+    include!("prompt_cache_and_timeseries/mod.rs");
+}
+
+pub(crate) use internal::*;
